Agencia OK uso de BigNumeral em Conta e começo de Transação

+++ src/main/java/acc/br/accenturebank/dto/ContaDTO.java
@@ -2,10 +2,12 @@ package acc.br.accenturebank.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class ContaDTO {
     private String numero;
-    private double saldo;
+    private BigDecimal saldo;
     private boolean ativa;
     private boolean pixAtivo;
     private String chavePix;
