+++ src/main/java/acc/br/accenturebank/dto/TransferenciaRequest.java
@@ -2,11 +2,13 @@ package acc.br.accenturebank.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class TransferenciaRequest {
     private Long idContaOrigem;
     private String numeroContaDestino;
-    private float valor;
+    private BigDecimal valor;
 
 
 }
